Add: edep data, Edit: data restructuring, Remove: HitData operator=
Major restructuring, in which HitData derived classes have static functions to let the SD know whether a step is to be ignored, how each step becomes a hit, and how all hits can be converted to RootData. SDs can now be abstracted by transforming into template classes with a generic HitData type.

Note that the operator= was removed for DetectorHit, HitData, and HitData's derived classes as it was found to be unnecessary, although this was not extensively tested.